Move password checker from core to auth
Password checker is used for checking when user change the password, it should belong to auth
